added some security to add, edit & delete product/colour views

+++ products/views.py
@@ -71,6 +71,11 @@ def product_detail(request, product_slug, category_slug):
 
 @login_required
 def add_product(request):
+
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have the correct permissions to complete this action.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST)
         if form.is_valid():
@@ -90,18 +95,25 @@ def add_product(request):
 
 @login_required
 def edit_product(request, product_id):
+
     product = get_object_or_404(Product, pk=product_id)
+
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have the correct permissions to complete this action.')
+        return redirect(reverse('product_detail',
+                                args=[ product.category.slug,
+                                        product.slug,
+                                ],))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, instance=product)
         if form.is_valid():
             form.save()
             messages.success(request, f'{product.product_name} was updated')
             return redirect(reverse('product_detail', 
-                                args=[ product.category.slug,
-                                        product.slug,
-                                    ],
-                                )
-                            )
+                                    args=[ product.category.slug,
+                                            product.slug,
+                                    ],))
         else:
             messages.error(request, f'Failed to update {product.product_name}. Please ensure the form is valid.')
     else:
@@ -116,7 +128,16 @@ def edit_product(request, product_id):
 
 @login_required
 def delete_product(request, product_id):
+    
     product = get_object_or_404(Product, pk=product_id)
+
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have the correct permissions to complete this action.')
+        return redirect(reverse('product_detail', 
+                                args=[ product.category.slug,
+                                        product.slug,
+                                ],))
+
     product.delete()
     messages.success(request, f'{product.product_name} was deleted.')
     return redirect(reverse("products"))
@@ -124,6 +145,10 @@ def delete_product(request, product_id):
 
 @login_required
 def add_colour(request):
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have the correct permissions to complete this action.')
+        return redirect(reverse('products'))
+
     if request.method == 'POST':
         form = ColourForm(request.POST)
         if form.is_valid():
@@ -145,6 +170,14 @@ def add_colour(request):
 def edit_colour(request, colour_id):
     colour = get_object_or_404(Colour, pk=colour_id)
     product  = get_object_or_404(Product, pk=colour.product.id)
+
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have the correct permissions to complete this action.')
+        return redirect(reverse('product_detail', 
+                                args=[ product.category.slug,
+                                        product.slug,
+                                    ],))
+
     if request.method == 'POST':
         form = ColourForm(request.POST, instance=colour)
         if form.is_valid():
@@ -153,9 +186,7 @@ def edit_colour(request, colour_id):
             return redirect(reverse('product_detail', 
                                 args=[ product.category.slug,
                                         product.slug,
-                                    ],
-                                )
-                            )
+                                    ],))
         else:
             messages.error(request, f'Failed to update {product.product_name} - {colour.colour}. Please ensure the form is valid.')
     else:
@@ -172,8 +203,17 @@ def edit_colour(request, colour_id):
 
 @login_required
 def delete_colour(request, colour_id):
+
     colour = get_object_or_404(Colour, pk=colour_id)
     product  = get_object_or_404(Product, pk=colour.product.id)
+
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have the correct permissions to complete this action.')
+        return redirect(reverse('product_detail', 
+                                args=[ product.category.slug,
+                                        product.slug,
+                                ],))
+
     colour.delete()
     messages.success(request, f'{product.product_name} - {colour.colour} was deleted.')
     return redirect(reverse("products"))
